move retrieve_headers() outside of Chain into utils.rs
