+++ chain/chain/src/lib.rs
@@ -7,7 +7,7 @@ pub use doomslug::{Doomslug, DoomslugBlockProductionReadiness, DoomslugThreshold
 pub use lightclient::{create_light_client_block_view, get_epoch_block_producers_view};
 pub use near_chain_primitives::{self, Error};
 pub use near_primitives::receipt::ReceiptResult;
-pub use store::utils::{get_chunk_clone_from_header, get_incoming_receipts_for_shard};
+pub use store::utils::{get_chunk_clone_from_header, get_incoming_receipts_for_shard, retrieve_headers};
 pub use store::{
     ChainStore, ChainStoreAccess, ChainStoreUpdate, LatestWitnessesInfo, MerkleProofAccess,
     ReceiptFilter,
