+++ chain/client/src/sync/header.rs
@@ -393,7 +393,7 @@ mod test {
     use near_async::time::{Clock, Duration, FakeClock, Utc};
     use near_chain::test_utils::{process_block_sync, setup, setup_with_tx_validity_period};
     use near_chain::types::Tip;
-    use near_chain::{BlockProcessingArtifact, Provenance};
+    use near_chain::{retrieve_headers, BlockProcessingArtifact, Provenance};
     use near_chain_configs::MutableConfigValue;
     use near_client_primitives::types::SyncStatus;
     use near_crypto::{KeyType, PublicKey};
@@ -873,7 +873,9 @@ mod test {
             match message {
                 NetworkRequests::BlockHeadersRequest { hashes, peer_id } => {
                     assert_eq!(peer_id, peer1.peer_info.id);
-                    let headers = chain2.retrieve_headers(hashes, MAX_BLOCK_HEADERS, None).unwrap();
+                    let headers =
+                        retrieve_headers(chain2.chain_store(), hashes, MAX_BLOCK_HEADERS, None)
+                            .unwrap();
                     assert!(!headers.is_empty(), "No headers were returned");
                     match chain.sync_block_headers(headers, &mut Vec::new()) {
                         Ok(_) => {}
